feat(use-case): add remove-book validation for invalid id

+++ src/use-cases/remove-book/remove-book.ts
@@ -1,4 +1,5 @@
 import { BooksRepository } from '@/repositories/books-repository'
+import { ResourceNotFoundError } from '../@errors/resource-not-found-error'
 
 interface RemoveBookUseCaseRequest {
 	id: string
@@ -8,6 +9,12 @@ export class RemoveBookUseCase {
 	constructor(private booksRepository: BooksRepository) {}
 
 	async use(data: RemoveBookUseCaseRequest) {
+		const checkBook = await this.booksRepository.findById(data.id)
+
+		if (!checkBook) {
+			throw new ResourceNotFoundError()
+		}
+
 		await this.booksRepository.remove(data.id)
 	}
 }
